Login, Statistics and a not so well working stat tracker.
Theres still a problem in the code which ill have to find a work around to. But for now. The update adds a Login feature and handling to that. Session start so the user stays logged in. Statistics window which should update everytime a player gets a bingo. BUUUUT its not working properly currently.